+++ testing/i18n-basic/es_MX/basic.ftl
